[fix] Fix authors colors not getting active
After a period of inactivity, the editor gets reloaded. When that
happen, the information about the authors' colors is lost. To fix this,
the teksto sends a message to etherpad telling to get updated data.
In this commit, we respond to that message and request the necessary
data from teksto.